% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_raw.R
\name{import_raw}
\alias{import_raw}
\title{import_raw
Import raw data from MaxQuant into a SQLite database}
\usage{
import_raw(peptides_file, proteins_file, db_file)
}
\arguments{
\item{peptides_file}{A character string of the path to the peptides file}

\item{proteins_file}{A character string of the path to the proteins file}

\item{db_file}{A character string of the path to the SQLite database file}
}
\description{
import_raw
Import raw data from MaxQuant into a SQLite database
}
