% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaPlot.R
\name{pcaPlot}
\alias{pcaPlot}
\title{pcaPlot
Generate a PCA plot for visual inspection of clusters}
\usage{
pcaPlot(data, group)
}
\arguments{
\item{data}{A data frame containing the data to be plotted. All columns will be passed to `prcomp` and should be numeric.}

\item{group}{A character vector containing the group information}
}
\value{
A ggplot object
}
\description{
pcaPlot
Generate a PCA plot for visual inspection of clusters
}
