% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_raw.R
\name{import_raw}
\alias{import_raw}
\title{Import raw data}
\usage{
import_raw(peptides_file, proteins_file, db_file)
}
\arguments{
\item{peptides_file}{A character string of the path to the peptides file}

\item{proteins_file}{A character string of the path to the proteins file}

\item{db_file}{A character string of the path to the SQLite database}
}
\description{
This function imports raw data from MaxQuant output files into an SQLite database
}
